update swig bindings and wrapper

+++ src/omega_edit/src/impl_/data_def.hpp
@@ -36,7 +36,7 @@ union omega_data_t {
 static_assert(8 == sizeof(omega_data_t), "size of omega_data_t is expected to be 8 bytes");
 
 inline omega_byte_t *omega_data_get_data(omega_data_t *data_ptr, int64_t capacity) {
-    return (capacity < sizeof(omega_data_t)) ? data_ptr->sm_bytes : data_ptr->bytes_ptr.get();
+    return (capacity < static_cast<int64_t>(sizeof(omega_data_t))) ? data_ptr->sm_bytes : data_ptr->bytes_ptr.get();
 }
 
 #endif//OMEGA_EDIT_DATA_DEF_HPP
